Refactor: Removed duplicate or redundant files
- Delete duplicate modal component files from ui/ folder:
  * ModalBody.tsx, ModalCloseButton.tsx, ModalContext.tsx
  * ModalFooter.tsx, ModalHeader.tsx, ModalOverlay.tsx, ModalPanel.tsx
  * modals.ts barrel export

- Make ModalCloseButton standalone by inlining button implementation
  * Remove dependency on ui/ModalCloseButton
  * Add size and variant styling directly in modal/ModalCloseButton

- Extract ModalSize and ModalAnimation types to ModalContainer
  * Move type definitions from ui/Modal.tsx to modal/ModalContainer.tsx
  * Update ModalContext and ModalPanel to import from ModalContainer
  * Remove dependency on old ui/Modal.tsx file

- Update imports to use modal folder components exclusively
  * ModalHeader now imports ModalCloseButton from local folder
  * All modal components are now self-contained within modal/ directory

This completes the modal system consolidation by removing the old duplicate
components and ensuring all modal-related code is properly organized within
the modal/ folder.
