--- src/app/components/ui/modals.ts
@@ -1,13 +1,0 @@
-// Modal components - export all from one place for convenience
-// Use this when you need multiple modal components together
-
-export { XIcon, type XIconProps } from "./icons/XIcon";
-export { Modal, type ModalProps, type ModalSize } from "./Modal";
-export { ModalBody, type ModalBodyProps } from "./ModalBody";
-export {
-  ModalCloseButton,
-  type ModalCloseButtonProps,
-} from "./ModalCloseButton";
-export { type ModalContextValue, useModalContext } from "./ModalContext";
-export { ModalFooter, type ModalFooterProps } from "./ModalFooter";
-export { ModalHeader, type ModalHeaderProps } from "./ModalHeader";
